nv-public: p3737: Disable UFSHCI
The UFSHCI controller is not supported for Tegra234 Jetson platforms and
so disable this controller. Note that this change makes no difference
because the Tegra UFS driver is never automatically loaded based
device-tree for Jetson platforms. The Tegra UFS driver does not populate
the MODULE_DEVICE_TABLE() macro and so the driver has to be manually
loaded.

Change-Id: Ifb3a588b9bbd08a71bad9a71aa59e8b0e0fc038a
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3051223
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>