t23x: nv-public: Fix serial aliases
When booting Linux with GRUB, the serial console does not show the
kernel boot messages as expected. By default GRUB does not add the
'console' kernel parameter and relies on device-tree to configure the
default serial port. Device-tree configures the default serial port by
setting the property 'stdout-path=serial0:115200n8' where 'serial0' is
an alises to one of the devices serial ports. The default serial port
for Tegra234 devices is the TCU0 and so 'serial0' should be mapped to
this interface. However, the 'serial0' is being updated to be mapped to
a different UART.

Fix this by removing the additional 'serialX' aliases in the base
overlay file and add any alises that are not already defined to the
appropriate board file. This does change the mapping of some aliases but
this aligns the boards with the aliases as they are defined upstream.

Bug 4264560

Change-Id: Icf6bdb1e7d5c9abf5bdbf5378b4ed8122910b507
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3072368
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-by: Paritosh Dixit <paritoshd@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>