t23x: nv-public: Remove unused clock and reset names
Clock and reset names are being removed from upstream device-tree for
devices that only have a single clock or reset. The QSPI driver does not
use the 'reset-names' property and so this can be removed for these
devices.

UART devices may use either the 8250 serial driver or the Tegra HS
serial driver. The default is the 8250 driver. When the Tegra HS serial
driver is used, the 'reset-names' property is required because the
driver specifically uses the reset name although there is only one. The
clock-names and reset-names for the 0x3100000 UART can be removed from
the base overlay file because the reset-names is correctly specified
in the files where the compatibility string for the Tegra HS serial
driver is set.

Bug 4037899

Change-Id: I501cd36609824e5703a6b756fc5f5389dd8d2368
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3074829
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>