t23x: overlay: add no-map to vpr carveout
The VPR carveout has special security constraints. It must not be
read by the CPU. Add "no-map" property to this region to avoid
issues.

Currently linux-next builds are failing to boot because we hit
an error related to this mapping.

[    0.000000] Call trace:
[    0.000000]  numa_memblks_init+0x28c/0x32c
[    0.000000]  numa_init+0x48/0x218
[    0.000000]  arch_numa_init+0x48/0x84
[    0.000000]  bootmem_init+0x6c/0x174
[    0.000000]  setup_arch+0x23c/0x5fc
[    0.000000]  start_kernel+0x68/0x710
[    0.000000]  __primary_switched+0x80/0x88
[    0.000000] Code: a8c77bfd d50323bf d65f03c0 f9800051 (c85f7c46)

Addition of no-map resolves the issue.

Bug 4749580

Change-Id: I16cd52d5ce969d3af5e1814c51e061e2d584cc22
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3208483
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>